[android] fix getting version errors
When adding the ot-br-posix to master branch of the manifest, the
Presubmit build generates an error on branch git_master on target
aosp_crosshatch-userdebug. The error shows that the ot-br-posix source
code is not under a git repo. So that we can't use the `git describe`
command to get the version string.

This CL checks if the source code is under a git repo, if not, retuns
a string "Unknown" as the version string.

Bug: 203492771
Test: build
Change-Id: I52aaf37e4ef4e7b061332140d18c0fe45fe3bd40