[logging] use program name for log id (#1953)
On Android, we use the "ot-daemon" name for the runnable program
and we want the syslog id be updated to "ot-daemon" as well.
Instead of hardcoding the syslog id, this commit derive the id
from the program name so it works for both linux and Android.

(cherry picked from commit 0de4d53e334822d4364ba9f18ff959057a74d0f9)
Change-Id: I9c746f82801c7b56d115e056059ed0113fc1defd