[NAT64] implement otPlatInfraIfDiscoverNat64Prefix
This CL implements otPlatInfraIfDiscoverNat64Prefix. On Android
platform, this API doesn't do any proactive discovery but it notifies
the completion of NAT64 prefix discovery so OT core can be notified to
re-evaluate the NAT64 state.

The implementation currently assumes that there is no NAT64 prefix
on the AIL and notifies the OpenThread stack accordingly.

Bug: 357479886
Test: atest ThreadNetworkIntegrationTests

Change-Id: Ieaff8427a77495e03bd33842c25961f7b5357515