Change OtDaemonState ephemeralKeyExpiryMillis to ephemeralKeyLifetimeMillis
The ephemeralKeyExpiryMillis is from system_clock which is not
monotonic, if user adjusted system time, the ephemeral key expiry time
calculated by Thread system server may become wrong. This change uses
steady_clock instead of system_clock since steady_clock is monotonic.

The steady_clock::now().time_since_epoch() is not related to the Unix epoch,
so this change passes the remainging lifetime in ms of ephemeral key to
the system service to calculate the expiry time.

Bug: 372598146
Test: atest CtsThreadNetworkTestCases

Change-Id: I08ec55a8d311642722f0629768c122877d59a8ad