[Thread] avoid disruption when joining the same network twice
The current join() implementation will tear down the existing network
and re-join the same network if this device has already joined the target
network. This will cause connectivity discruptions as well as additional
delay for this device to re-join (especially when this device is the
first router / leader device).

This commits optimize this case by checking the equality of the Active
Dataset and returns success directly if this device has already joined
the exact same network.

Bug: 374037595
Test: atest ThreadNetworkIntegrationTests:android.net.thread.ThreadIntegrationTest#joinNetwork_joinTheSameNetworkTwice_noDeviceRoleChanges
Change-Id: I3a3443b0892e7a11464ca86dd1c6b53511a64f75