Add AospThreadHost with SetConfiguration method
This CL is a breakdown of a larger CL. This CL introduces an interface
AospThreadHost to provide a unified Thread control interface for both
NCP & RCP cases to implement Android specific Thread APIs (otdaemon
binder APIs).

This CL adds two methods in `AospThreadHost`: `SetConfiguration` and
`GetConfiguration`. And it replaces the implementation of
`setConfiguration` with a call to `AospThreadHost::SetConfiguration`.

Bug: 377175558
Test: Build mainline module. "m apps_only dist".
Test: Existing unit tests can still pass.
Change-Id: Ic330b27f731dbdbdfb1964426adce485d014351d