[Thread] allow disabling SRP server wait for border routing
It's default enabled that it waits for border routing establishment
before starting the SRP server. But there are also use cases where
the user may want to know what are the devices in the Thread network
before actually reach to the devices, or it doesn't depend on border
routing to reach to a Thread device. Given that, this commit adds the
support of starting SRP server without waiting for border routing by
leveraging a new configuration parameter:
srpServerWaitForBorderRoutingEnabled.

Test: see aosp/3391245
Bug: 378587036
Change-Id: I2faecf06b3495591196753f85bafb292676bd6aa