Add //apex_available:platform to ot_daemon_service_fuzzer
Soong was accidentally allowing a corner case where modules that
were not supposed to be available to the platform (due to
apex_available being set to an apex and not including
"//apex_available:platform") were being made available to the
platform as long as no apex modules depended on them.  Add
"//apex_available:platform" explicitly to ot_daemon_service_fuzzer
so that it is available to the haiku-presubmit targets.

Bug: 394664568
Test: ABTD
Change-Id: Ibea39bba4d60d9ae307c69c68dfc007b410d761b